refactor(terminal-workspace): extract command run model
